refactor(html5): deduplicate insert-pages upload and blank-page code
The insert-pages feature landed with three copies of new code:

- the content-to-File helpers (size gate, base64 decoding, content
  materialization) were copied from the plugin upload command manager
  into the insert-pages command manager; they now live in a shared
  content-to-file module used by both.
- the upload-token/multipart insert flow existed in both the native
  toolbar service and the plugin command service; the plugin manager
  now calls the toolbar's insertPagesUpload and the duplicate service
  is gone. Failures keep being logged under the plugin logCode by the
  manager's existing catch.
- blank pages had two contradictory mechanisms: the toolbar fetched
  the server-hosted blank.pdf while the plugin path synthesized a
  1920x1080 PNG on a canvas. Both now insert the same blank.pdf
  through the regular conversion pipeline.
