+++ bigbluebutton-html5/imports/ui/components/plugins-engine/server-commands/presentation/upload/manager.tsx
@@ -1,7 +1,6 @@
 import { useEffect } from 'react';
 import {
   UploadPresentationCommandArguments,
-  UploadPresentationContent,
 } from 'bigbluebutton-html-plugin-sdk/dist/cjs/server-commands/presentation/types';
 import {
   PresentationCommandsEnum,
@@ -10,58 +9,7 @@ import { uniqueId } from '/imports/utils/string-utils';
 import PresentationUploaderService from '/imports/ui/components/actions-bar/media-area/media-sharing/presentation/service';
 import logger from '/imports/startup/client/logger';
 import useCurrentUser from '/imports/ui/core/hooks/useCurrentUser';
-
-const getMaxBytes = (): number => window.meetingClientSettings?.public?.presentation
-  ?.mirroredFromBBBCore?.uploadSizeMax ?? 30_000_000;
-
-const assertSize = (bytes: number) => {
-  const maxBytes = getMaxBytes();
-  if (bytes > maxBytes) {
-    throw new Error(`Presentation payload exceeds the maximum allowed size of ${maxBytes} bytes.`);
-  }
-};
-
-const decodeBase64ToFile = (
-  rawBase64: string,
-  mimeType: string,
-  presentationName: string,
-): File => {
-  // base64 encodes 3 bytes as 4 chars; this gives the upper-bound decoded size.
-  assertSize(Math.ceil((rawBase64.length * 3) / 4));
-  const bytes = Uint8Array.from(atob(rawBase64), (c) => c.charCodeAt(0));
-  return new File([new Blob([bytes], { type: mimeType })], presentationName, { type: mimeType });
-};
-
-const toFile = async (
-  content: UploadPresentationContent,
-  mimeType: string,
-  name?: string,
-): Promise<File> => {
-  const ext = mimeType.split('/')[1] || 'pdf';
-  const presentationName = name || `Plugin_Presentation.${ext}`;
-
-  if ('file' in content) {
-    assertSize(content.file.size);
-    return content.file;
-  }
-
-  if ('blob' in content) {
-    assertSize(content.blob.size);
-    return new File([content.blob], presentationName, { type: mimeType });
-  }
-
-  if ('dataUrl' in content) {
-    const match = content.dataUrl.match(/^data:[^;]+;base64,(.+)$/);
-    if (!match) throw new Error('Invalid or non-base64 dataURL.');
-    return decodeBase64ToFile(match[1], mimeType, presentationName);
-  }
-
-  if ('base64' in content) {
-    return decodeBase64ToFile(content.base64, mimeType, presentationName);
-  }
-
-  throw new Error('Object type not supported.');
-};
+import { toFile } from '../content-to-file';
 
 const isValidUploadEvent = (event: CustomEvent<UploadPresentationCommandArguments>) => {
   if (
